# Remember column indices between different list items